Revert "Faster implementation of backend sort"
