+++ Cargo.toml
@@ -4,7 +4,10 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+humantime = "2.1.0"
 itertools = "0.13.0"
+notoize = "2.13.0"
+regex = "1.10.6"
 reqwest = { version = "0.12.5", features = ["blocking"] }
 serde = { version = "1.0.204", features = ["derive"] }
 serde_json = "1.0.122"
