--- src/tests.rs
@@ -1,1 +1,0 @@
-use crate::old_main::{self, tones};
